Simplify WASM extension organization
