+++ apollo-router/Cargo.toml
@@ -250,8 +250,8 @@ tracing = "0.1.40"
 tracing-core = "0.1.32"
 tracing-futures = { version = "0.2.5", features = ["futures-03"] }
 tracing-subscriber = { version = "0.3.20", features = ["env-filter", "json"] }
-wasmtime = { version = "47.0.2", features = ["async", "component-model"] }
-wasmtime-wasi = "47.0.2"
+wasmtime = { version = "47.0.2", default-features = false, features = ["async", "cache", "component-model", "cranelift", "parallel-compilation", "runtime", "std"] }
+wasmtime-wasi = { version = "47.0.2", default-features = false, features = ["p2"] }
 url = { version = "2.5.4", features = ["serde"] }
 urlencoding = "2.1.3"
 uuid = { version = "1.9.1", features = ["serde", "v4"] }
