+++ examples/wasm-extensions/java-header/wrapper.js
@@ -1,27 +1,4 @@
 import { headerValue } from "./target/javascript/plugin.js";
+import { headerPlugin } from "../tooling/header-plugin.js";
 
-export const hooks = {
-  handle() {
-    return {
-      tag: "proceed",
-      val: {
-        headers: [
-          {
-            tag: "set",
-            val: { name: "x-wasm-java", values: [headerValue()] },
-          },
-        ],
-        context: [
-          {
-            tag: "set",
-            val: {
-              name: "wasm.java",
-              value: JSON.stringify({ language: "java" }),
-            },
-          },
-        ],
-        body: undefined,
-      },
-    };
-  },
-};
+export const hooks = headerPlugin("java", headerValue);
