+++ examples/wasm-extensions/node-header/.gitignore
@@ -1,2 +1,2 @@
-/node_modules/
+/target/
 /plugin.wasm
