--- examples/wasm-extensions/node-header/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "router-wasm-node-header",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "jco componentize plugin.js --wit ../../../apollo-router/wit/router-plugin --world-name router-plugin --disable all -o plugin.wasm"
-  },
-  "devDependencies": {
-    "@bytecodealliance/componentize-js": "0.22.0",
-    "@bytecodealliance/jco": "1.26.1",
-    "esbuild": "0.25.10"
-  }
-}
