+++ examples/wasm-extensions/node-header/plugin.js
@@ -1,25 +1,3 @@
-export const hooks = {
-  handle(event) {
-    return {
-      tag: "proceed",
-      val: {
-        headers: [
-          {
-            tag: "set",
-            val: { name: "x-wasm-node", values: ["active"] },
-          },
-        ],
-        context: [
-          {
-            tag: "set",
-            val: {
-              name: "wasm.node",
-              value: JSON.stringify({ language: "node" }),
-            },
-          },
-        ],
-        body: undefined,
-      },
-    };
-  },
-};
+import { headerPlugin } from "../tooling/header-plugin.js";
+
+export const hooks = headerPlugin("node");
