+++ examples/wasm-extensions/scala-header/build.sh
@@ -19,13 +19,13 @@ scala-cli --power package \
   --force \
   Plugin.scala \
   --output target/plugin.js
-../node-header/node_modules/.bin/esbuild wrapper.js \
+../node_modules/.bin/esbuild wrapper.js \
   --bundle \
   --format=esm \
   --platform=neutral \
   --target=es2022 \
   --outfile=target/plugin.bundle.js
-../node-header/node_modules/.bin/jco componentize \
+../node_modules/.bin/jco componentize \
   target/plugin.bundle.js \
   --wit ../../../apollo-router/wit/router-plugin \
   --world-name router-plugin \
