+++ examples/wasm-extensions/scala-header/wrapper.js
@@ -1,27 +1,4 @@
 import { headerValue } from "./target/plugin.js";
+import { headerPlugin } from "../tooling/header-plugin.js";
 
-export const hooks = {
-  handle() {
-    return {
-      tag: "proceed",
-      val: {
-        headers: [
-          {
-            tag: "set",
-            val: { name: "x-wasm-scala", values: [headerValue()] },
-          },
-        ],
-        context: [
-          {
-            tag: "set",
-            val: {
-              name: "wasm.scala",
-              value: JSON.stringify({ language: "scala" }),
-            },
-          },
-        ],
-        body: undefined,
-      },
-    };
-  },
-};
+export const hooks = headerPlugin("scala", headerValue);
